omni-scan/backend/requirements.txt
# IMPORTANT: NumPy constraint must be first to ensure compatibility
numpy<2.0

# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2
websockets==12.0

# Database
supabase==2.10.0
asyncpg==0.30.0

# OCR - Version compatible avec PaddleOCR
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
# Supprimé opencv-python-headless car PaddleOCR installe ses propres versions d'OpenCV
PyPDF2==3.0.1

# OCR Léger - Modèles optimisés VPS
easyocr==1.7.0
paddlepaddle==2.6.2
paddleocr==2.7.3
psutil==5.9.6

# AI - Mise à jour critique
openai==1.97.1
langchain==0.3.27
tiktoken==0.8.0

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
aiofiles==24.1.0
redis==5.0.7
stripe==10.2.0
pycryptodome==3.20.0
python-dotenv==1.0.1
PyJWT==2.8.0
jinja2==3.1.4
werkzeug==3.0.3
rich==13.7.1

# Email validation
email-validator==2.2.0

# Dev
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.8.1

# Logging
python-json-logger==3.2.1

# Export
reportlab==4.2.5
openpyxl==3.1.5
pandas==1.5.3  # Compatible avec numpy<2.0
